requirements.txt
torch>=1.10
torchvision>=0.11
timm>=0.5.4
opencv-python>=4.5
albumentations>=1.0.0
scikit-learn>=0.24
pyyaml
pandas
numpy
Pillow
matplotlib
optuna>=3.0
